etl/requirements.txt
kafka-python==2.0.2
aiokafka==0.8.1
clickhouse-driver==0.2.6
pydantic==1.9.1
python-dotenv==1.0
uvicorn==0.23.2
gunicorn==21.2.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
httptools==0.6.0
orjson==3.9.2